+++ jvm/src/main/scala/poolmate/Store.scala
@@ -325,7 +325,7 @@ final class Store(conf: Config,
   def listHeaters(): List[Heater] =
     DB readOnly { implicit session =>
       sql"select * from heater order by installed desc"
-        .map(rs => Heater(rs.long("id"), rs.long("pool_id"), rs.int("installed"), rs.string("model"), rs.double("cost")))
+        .map(rs => Heater(rs.long("id"), rs.long("pool_id"), rs.int("installed"), rs.string("model"), rs.int("cost")))
         .list()
     }
