+++ shared/src/main/scala/poolmate/Entity.scala
@@ -140,7 +140,7 @@ final case class Heater(id: Long = 0,
                         poolId: Long = 0,
                         installed: Int = 0,
                         model: String = "",
-                        cost: Double = 0.0) extends Entity:
+                        cost: Int = 0) extends Entity:
   def display = installed.toString
 
 final case class HeaterSetting(id: Long = 0,
